refactor: common 파일 재분류
